scripts/bin/auto-install: refactor check_drive
Instead of doing a messy recursive loop, instead read the contents of
the watch file into a variable and use that variable, and check if that
variable is empty or not. If the variable is empty, continue with the
loop; we don't need to recursively call the function because we already
have a busy do loop.

We'll also force-copy the log to ensure we always get the log onto the
drive.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>