scripts/bin/auto-install: make curl a function
Instead of reproducing a curl call in each unique function, pull it out
to a separate rest_call function which interacts with snapd via its REST
API. This allows for extending the behavior in a simple way if e.g. we
want more introspection into success, failure, other actions, etc.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>