auto-mount: bail if target dir is a mountpoint
Somehow, the symlinks in /dev/disk/by-uuid have the potential of being
created multiple times in rapid succession. Because this symlink has the
same name both times, we will end up attempting to mount the same
partition to the same location twice. The first time will succeed just
fine and auto-install will proceed to install all of the snaps. After a
short amount of time, however, try_mount() will attempt to resolve the
pending mount of the already mounted partition to the already mounted-on
target directory. Because the target directory is already a mountpoint,
the mount operation will fail.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>